requirements_rag.txt
chromadb==0.4.15
sentence-transformers==2.7.0
transformers==4.41.0
gradio>=4.0.0
google-generativeai>=0.3.0
openai>=1.0.0
python-dotenv
tqdm
